Security update + Password reset
Password is now hashed and saved in the database to add some sort of
security. For now, this is just hashing but in a future build, this will
be upgraded to support salting as well to make password saving even more
secure. Users can now also reset their password if forgotten. Finally,
made some general UI updates to the homepage and dashboard.